fix(base_url): fix fixed env variable name cannot fit with nextjs, pass variable instead
A BASE_URL environment variable was expected in the environment, however in addition to not seeing
the expected errors when not available, frameworks like nextjs require custom names like
NEXT_PUBLIC_BASE_URL for client code. Instead in this change the baseURL is now passed in the
constructor.

BREAKING CHANGE: The base URL now needs to be passed in the constructor instead of handled in the
.env